V3 Alpha Code to Support Ambient Remote Sensors
 have made some early progress in incorporating Ambient WH31B remote sensors into V3 version.

This V3 version picks up the N number of Ambient remote sensors that are reporting from Ambient Network API and creates separate SmartThings devices and Tiles for displaying and handling events from their individual temperature and humidity displays.
The remote sensors networkID are identified in the V3 version by using integer 1-8 which is appended to each network core name and is the dip switch setting 1-8 on the remote sensor.
The V3 app has been completely redesigned to recognize remote sensors and allows one to provide a short location name for identification of each remote sensor in the SmartThings mobile client.
The V3 install now names the tile for the Ambient Weather console the descriptive same as what it is named in the Ambient Network Dashboard. If one wants to change that name, they must change it on the Ambient Network dashboard settings page.
There is still a limitation of allowing only ONE weather console to be integrated to SmartThings per Ambient Weather Station SmartApp, which is usually what most people own.