Bug Fix 6.2.1
A Hubitat library name cannot have a "-" in the filename when it is extracted as a zip file. The bundle in zip format silently removes the offending hyphen in the library and does not generate an error message.  This release removes the "-" in the library name and the #includes in the source code files.